+++ scripts/validate_mac_mot_v2_checkpoint.py
@@ -29,7 +29,7 @@ def main() -> None:
         chunk_horizon=48,
         expert_hidden_dim=args.expert_hidden_dim,
         device=args.device,
-        dtype=torch.bfloat16,
+        dtype=torch.float32,
     )
     trainable = configure_trainable_parameters(model, "critic")
     old_modules = [
@@ -69,14 +69,14 @@ def main() -> None:
             2,
             model.txt_in.in_features,
             device=device,
-            dtype=torch.bfloat16,
+            dtype=torch.float32,
         )
         current = torch.randn(
-            batch, 12 * 24, model.in_channels, device=device, dtype=torch.bfloat16
+            batch, 12 * 24, model.in_channels, device=device, dtype=torch.float32
         )
-        state = torch.randn(batch, 1, model.state_dim, device=device, dtype=torch.bfloat16)
+        state = torch.randn(batch, 1, model.state_dim, device=device, dtype=torch.float32)
         action = torch.randn(
-            batch, model.chunk_horizon, model.action_dim, device=device, dtype=torch.bfloat16
+            batch, model.chunk_horizon, model.action_dim, device=device, dtype=torch.float32
         )
         context_ids = text_position_ids(batch, context.shape[1], device)
         current_ids = image_position_ids(
@@ -104,9 +104,7 @@ def main() -> None:
             clean_action=action,
             context_mask=torch.ones(batch, context.shape[1], device=device, dtype=torch.bool),
         )
-        # The rollout path intentionally pairs BF16 frozen-FLUX caches with an
-        # FP32 target Value expert. Exercise that exact mixed-dtype boundary,
-        # not only the two online BF16 experts.
+        # Target Value uses the same FP32 computation as the online experts.
         target_value = ValueExpertEMA(
             model.value_expert,
             device=device,
